Move FakeOperation to test.helpers
