+++ tests/operations/test_base.py
@@ -1,37 +1,18 @@
-from dataclasses import asdict, dataclass, field
-from typing import Any, Dict, Optional
 from unittest import TestCase
 from unittest.mock import Mock, patch
 
 import pytest
 
-from qwikstart.base_context import BaseContext, DictContext
+from qwikstart.base_context import DictContext
 from qwikstart.operations import base
 
 from .. import helpers
 
 
-@dataclass(frozen=True)
-class ContextWithDict(BaseContext):
-    template_variables: Dict[str, Any] = field(default_factory=dict)
-
-
-class FakeOperation(base.BaseOperation[ContextWithDict, DictContext]):
-    name: str = "fake-operation"
-
-    def __init__(self, **kwargs: Any) -> None:
-        super().__init__(**kwargs)
-        self.run_context: Optional[ContextWithDict] = None
-
-    def run(self, context: ContextWithDict) -> DictContext:
-        self.run_context = context
-        return asdict(context)
-
-
-class ErrorOperation(base.BaseOperation[ContextWithDict, DictContext]):
+class ErrorOperation(base.BaseOperation[helpers.ContextWithDict, DictContext]):
     name: str = "error-operation"
 
-    def run(self, context: ContextWithDict) -> DictContext:
+    def run(self, context: helpers.ContextWithDict) -> DictContext:
         raise Exception("Error raised for testing purposes")
 
 
@@ -40,14 +21,14 @@ class TestOperationHavingContextWithDict(TestCase):
         self.execution_context = helpers.get_execution_context()
 
     def test_default_template_variables(self) -> None:
-        operation = FakeOperation()
+        operation = helpers.FakeOperation()
         output = operation.execute({"execution_context": self.execution_context})
-        assert isinstance(operation.run_context, ContextWithDict)
+        assert isinstance(operation.run_context, helpers.ContextWithDict)
         assert operation.run_context.template_variables == {}
         assert output["template_variables"] == {}
 
     def test_template_variables_from_run_context(self) -> None:
-        operation = FakeOperation()
+        operation = helpers.FakeOperation()
         template_variables = {"some": "value"}
         output = operation.execute(
             {
@@ -55,13 +36,13 @@ class TestOperationHavingContextWithDict(TestCase):
                 "template_variables": template_variables,
             }
         )
-        assert isinstance(operation.run_context, ContextWithDict)
+        assert isinstance(operation.run_context, helpers.ContextWithDict)
         assert operation.run_context.template_variables == template_variables
         assert output["template_variables"] == template_variables
 
     def test_input_mapping(self) -> None:
         config = base.OperationConfig(input_mapping={"vars": "template_variables"})
-        operation = FakeOperation(config=config)
+        operation = helpers.FakeOperation(config=config)
         output = operation.execute(
             {"execution_context": self.execution_context, "vars": {"some": "value"}}
         )
@@ -71,7 +52,7 @@ class TestOperationHavingContextWithDict(TestCase):
         config = base.OperationConfig(
             output_mapping={"template_variables": "output_vars"}
         )
-        operation = FakeOperation(config=config)
+        operation = helpers.FakeOperation(config=config)
         output = operation.execute(
             {
                 "execution_context": self.execution_context,
@@ -81,7 +62,7 @@ class TestOperationHavingContextWithDict(TestCase):
         assert output["output_vars"] == {"some": "value"}
 
     def test_local_context_takes_precedence(self) -> None:
-        operation = FakeOperation(
+        operation = helpers.FakeOperation(
             local_context={"template_variables": {"same-key": "local-context"}}
         )
         output = operation.execute(
@@ -93,7 +74,9 @@ class TestOperationHavingContextWithDict(TestCase):
         assert output["template_variables"] == {"same-key": "local-context"}
 
     def test_nested_dictionary_merged_from_local_context(self) -> None:
-        operation = FakeOperation(local_context={"template_variables": {"b": 2}})
+        operation = helpers.FakeOperation(
+            local_context={"template_variables": {"b": 2}}
+        )
         output = operation.execute(
             {
                 "execution_context": self.execution_context,
@@ -103,17 +86,17 @@ class TestOperationHavingContextWithDict(TestCase):
         assert output["template_variables"] == {"a": 1, "b": 2}
 
     def test_repr(self) -> None:
-        operation = FakeOperation(description="Test Op")
+        operation = helpers.FakeOperation(description="Test Op")
         args = (
             "local_context={}, "
             f"config={base.OperationConfig()}, "
             "description=Test Op"
         )
-        assert repr(operation) == f"tests.operations.test_base.FakeOperation({args})"
+        assert repr(operation) == f"tests.helpers.FakeOperation({args})"
 
     @patch.object(base, "logger")
     def test_log_success_if_description_defined(self, logger: Mock) -> None:
-        operation = FakeOperation(description="Step 1")
+        operation = helpers.FakeOperation(description="Step 1")
         operation.execute({"execution_context": self.execution_context})
         logger.info.assert_called_once_with(f"Step 1: {base.SUCCESS_MARK}")
 
@@ -126,7 +109,7 @@ class TestOperationHavingContextWithDict(TestCase):
 
     @patch.object(base, "logger")
     def test_log_success_not_called_without_description(self, logger: Mock) -> None:
-        FakeOperation().execute({"execution_context": self.execution_context})
+        helpers.FakeOperation().execute({"execution_context": self.execution_context})
         logger.info.assert_not_called()
 
     @patch.object(base, "logger")
